change 'filterlist' to 'imagelist', use productmd formats
As an alternative to filtering out known not-image file types,
just import the list of known image file types from productmd
and filter out anything not in that when creating the list for
fedfind. Now call it 'imagelist' rather than 'filterlist'.

The advantage of this is it gives us the smallest possible list
and we don't have to worry about the filterlist file sizes
suddenly growing because a large number of some new type of
file starts showing up in the compose process (this would make
fedfind runs suddenly use a lot more bandwidth, but otherwise
wouldn't be terribly visible, which is kinda bad).

The disadvantage is it introduces a dependency on productmd
which we'd have to ensure we always account for on the systems
where create-filelist is run, and keep up to date so we don't
miss new image types.

At tibbs' request, make create-filelist work if productmd is
not available, to keep it generically useful: if productmd is
not available, the imagelist feature is disabled.