add a client-side filtering option (#16)
This adds a new feature. The user can specify a FILTEREXP in
the config. If specified, this is expected to be a regular
expression that quick-fedora-mirror and quick-fedora-hardlink
will use to filter the file lists they work with. The regex
is passed to `sed -i -r -e \,(expression),d` by q-f-m and
`re.compile(expression)` by q-f-h (after stripping any quotes,
in the latter case). (q-f-h.zsh does the same as q-f-m).

This allows you to maintain a partial mirror using q-f-m. Items
filtered out by the regex are treated as if they simply didn't
exist on the remote end. If any are present locally they'll be
treated as 'stale' files and removed unless skipdelete is set.
They will never be synchronized.