Use checksum (not mtime) to check if filelist changed
My mirror files are on an NFS share, which seems to mess with
the mtime check. That seems like a slightly fragile approach
anyhow, so why not just md5sum the list? This re-purposes the
fltimes array, which AFAICS was never used for anything at all,
as a checksums array and uses md5sum for the 'file list did not
change, bail out now' check. Also entirely removes the disabled
diffing code; having commented-out old stuff all over the place
is awful for readability, the commit history is the correct
place to memorialize old mistakes ;)