Major refactoring to support protected content
The client now supports protected content as marked by the updated
filelist generator.

Much code was hoisted out of the main program flow and into separate
functions.  Making sure this wasn't breaking the world was the reason
for the test suite.  More work is needed before we hit an actual
release.