Work around weird MM mapping for fedora-enchilada.
Mirrormanager wants to see checkins without the directory prefix, but
for the "fedora-enchilada" module, the rsync module is in "fedora" but
mirrormanager wants "fedora/linux" to be stripped.  So work up a
separate associative array to allow mirrormanager to differ.  It's just
a copy of the regular module->directory mapping, with the entry for
"fedora-enchilada" changed.